feat: Simplify game camera.
