+++ src/PixelDust.Game/Camera/PCamera.cs
@@ -1,35 +1,148 @@
 ﻿using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
 
 using PixelDust.Game.Managers;
 
+using System;
+
 namespace PixelDust.Game.Camera
 {
-    public abstract class PCamera(PScreenManager screenManager)
+    public sealed class PCamera
     {
-        public abstract Vector2 Position { get; set; }
-        public abstract float Rotation { get; set; }
-        public abstract float Zoom { get; set; }
-        public abstract float MinimumZoom { get; set; }
-        public abstract float MaximumZoom { get; set; }
-        public abstract Rectangle BoundingRectangle { get; }
-        public abstract Vector2 Origin { get; set; }
-        public abstract Vector2 Center { get; }
-        public PScreenManager ScreenManager => screenManager;
+        public Vector2 Position { get; set; }
+        public float Rotation { get; set; }
+        public Vector2 Origin { get; set; }
+        public Vector2 Center => this.Position + this.Origin;
+        public float Zoom
+        {
+            get => this._zoom;
+            set
+            {
+                if (value < this.MinimumZoom || value > this.MaximumZoom)
+                {
+                    throw new ArgumentException("Zoom must be between MinimumZoom and MaximumZoom");
+                }
 
-        public abstract void Move(Vector2 direction);
-        public abstract void Rotate(float deltaRadians);
-        public abstract void ZoomIn(float deltaZoom);
-        public abstract void ZoomOut(float deltaZoom);
-        public abstract void LookAt(Vector2 position);
+                this._zoom = value;
+            }
+        }
+        public float MinimumZoom
+        {
+            get => this._minimumZoom;
+            set
+            {
+                if (value < 0)
+                {
+                    throw new ArgumentException("MinimumZoom must be greater than zero");
+                }
 
-        public abstract Vector2 WorldToScreen(Vector2 worldPosition);
-        public abstract Vector2 ScreenToWorld(Vector2 screenPosition);
+                if (this.Zoom < value)
+                {
+                    this.Zoom = this.MinimumZoom;
+                }
 
-        public abstract Matrix GetViewMatrix();
-        public abstract Matrix GetInverseViewMatrix();
+                this._minimumZoom = value;
+            }
+        }
+        public float MaximumZoom
+        {
+            get => this._maximumZoom;
+            set
+            {
+                if (value < 0)
+                {
+                    throw new ArgumentException("MaximumZoom must be greater than zero");
+                }
 
-        public abstract BoundingFrustum GetBoundingFrustum();
-        public abstract ContainmentType Contains(Vector2 vector2);
-        public abstract ContainmentType Contains(Rectangle rectangle);
+                if (this.Zoom > value)
+                {
+                    this.Zoom = value;
+                }
+
+                this._maximumZoom = value;
+            }
+        }
+
+        private float _maximumZoom = float.MaxValue;
+        private float _minimumZoom;
+        private float _zoom;
+
+        private readonly PScreenManager _screenManager;
+
+        public PCamera(PScreenManager screenManager)
+        {
+            this._screenManager = screenManager;
+            this.Rotation = 0;
+            this.Zoom = 1;
+            this.Origin = new Vector2(screenManager.DefaultResolution.Width, screenManager.DefaultResolution.Height) / 2f;
+            this.Position = Vector2.Zero;
+        }
+
+        public void Move(Vector2 direction)
+        {
+            this.Position += Vector2.Transform(direction, Matrix.CreateRotationZ(-this.Rotation));
+        }
+
+        public void Rotate(float deltaRadians)
+        {
+            this.Rotation += deltaRadians;
+        }
+
+        public void ZoomIn(float deltaZoom)
+        {
+            ClampZoom(this.Zoom + deltaZoom);
+        }
+
+        public void ZoomOut(float deltaZoom)
+        {
+            ClampZoom(this.Zoom - deltaZoom);
+        }
+
+        private void ClampZoom(float value)
+        {
+            this.Zoom = value < this.MinimumZoom ? this.MinimumZoom : value > this.MaximumZoom ? this.MaximumZoom : value;
+        }
+
+        public void LookAt(Vector2 position)
+        {
+            this.Position = position - (new Vector2(this._screenManager.DefaultResolution.Width, this._screenManager.DefaultResolution.Height) / 2f);
+        }
+
+        public Vector2 WorldToScreen(float x, float y)
+        {
+            return WorldToScreen(new Vector2(x, y));
+        }
+
+        public Vector2 WorldToScreen(Vector2 worldPosition)
+        {
+            Viewport viewport = _screenManager.Viewport;
+            return Vector2.Transform(worldPosition + new Vector2(viewport.X, viewport.Y), GetViewMatrix());
+        }
+
+        public Vector2 ScreenToWorld(Vector2 screenPosition)
+        {
+            Viewport viewport = this._screenManager.Viewport;
+            return Vector2.Transform(screenPosition - new Vector2(viewport.X, viewport.Y),
+                   Matrix.Invert(GetViewMatrix()));
+        }
+
+        private Matrix GetVirtualViewMatrix()
+        {
+            return Matrix.CreateTranslation(new(-this.Position.X, this.Position.Y, 0.0f)) *
+                   Matrix.CreateTranslation(new(-this.Origin, 0.0f)) *
+                   Matrix.CreateRotationZ(this.Rotation) *
+                   Matrix.CreateScale(this.Zoom, this.Zoom, 1) *
+                   Matrix.CreateTranslation(new(this.Origin, 0.0f));
+        }
+
+        public Matrix GetViewMatrix()
+        {
+            return GetVirtualViewMatrix() * Matrix.Identity;
+        }
+
+        public Matrix GetInverseViewMatrix()
+        {
+            return Matrix.Invert(GetViewMatrix());
+        }
     }
 }
